pkg/mongo/mongo.go
package mongo

import (
	"context"

	"go.mongodb.org/mongo-driver/mongo"
	"go.mongodb.org/mongo-driver/mongo/options"
	"go.mongodb.org/mongo-driver/mongo/readpref"
)

// Config is a struct that contains the options for the mongodb connection
type Config struct {
	URI      string
	Database string
}

// Connect creates a connection to the mongodb database
func Connect(ctx context.Context, cfg Config) (*mongo.Database, error) {
	client, err := mongo.Connect(ctx, options.Client().ApplyURI(cfg.URI))
	if err != nil {
		return nil, err
	}
	if err := client.Ping(context.TODO(), readpref.Primary()); err != nil {
		return nil, err
	}
	return client.Database(cfg.Database), nil
}
